Add capacity manager functionality in swifttool. Remove other utility modules
